[2607-DEV-496] Step 1: extract LOS import phase-machine + API orchestration into useLosImport hook
